requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
flask_sqlalchemy==3.1.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lyricsgenius==3.0.1
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.2
psycopg2==2.9.9
plotly==5.19.0
pandas==2.2.1
unidecode==1.3.8